fix: don't allow ld-floxlib.so to bring in its own libc.so
Compiling ld-floxlib.so with its own RPATH or RUNPATH had the effect
of bringing in its own version of libc.so which would produce segfaults
when the library version didn't precisely match the one that had already
been brought in by the loader. This commit removes the RPATH or RUNPATH
entirely from the library so that it will use the version of libc.so
already loaded into the runtime.